more impl., ActionTargetManager, bug bix, ...
Finished more of the ActionTarget/Task...  hierarchy implementation.

Began initial coding/design of the concept of an ActionTargetManager class
whose job it is to manage the execution of "actions" with respect to tasks
and memoranda, e.g., to send an email reminder about a task whose deadline
is coming up.  Much more to go on this.

Bug fix in scanning of spec files re. what to do when the label (e.g.,
"type:") is followed by an empty value - i.e.:

-    components = spec_string.split(split_regex)
+    components = spec_string.split(split_regex, -1)

Added a big more of error or unexpected-situation handling.