keyword changes, strip out comments
Replaced the 'media' keyword with 'email' and 'calendar' keywords, so that
email addresses are designated separately from "calendar handles" (one or
more handles used to identify the user(s) for calendar entries).
Currently, only google-calendar will be targeted for this functionality
(i.e., adding an entry in the calendar based on contents of a spec file);
but it's possible other calendar systems accessible via an API will be added
later.

STodoSpec: Added 'stripped_of_comments'.  Refactored the initialization a
bit.