Prevent duplicate and missing handles.
Added the planned functionality to disallow specs that are missing a
"handle" field - i.e., in that case display a warning message to the user
and don't move the spec file to the 'post_init_spec_path'.  Also,
implemented saving all new handles encountered into a file that is
used to check if a new spec uses a handle that has already been used.  If
that situation (already used) is encountered, the user is warned of this
and the spec/target is not processed and the spec file is left as is.
Added a new class, FileBasedDataManager, to take care of the file-handling
aspects (create, write, and read handles file) of the
handle-duplicate-prevention functionality.