Started implementation of "ongoing" processing.
First, incomplete, attempt at "ongoing" processing - i.e., restoring items
(tasks, appointments, etc.) that have already been initialized and
performing needed ongoing actions with respect to these items (such as
sending a reminder for a task if the current date/time is close to that of
one of the tasks reminders).  This version, which is likely to be replaced
by one that uses serialization, simply reads the moved spec files and
creates the needed STodoTarget objects from them, just as the
"initial-processing" logic does and then it calls
perform_ongoing_processing on the STodoManager objects instead of
perform_initial_processing.