Switched to serialization of initialized specs...
Changed to use serialization (via Marshal::dump and saving the result to a
file and Marshal::load for restoration) for storing "STodoTarget"s and
retrieving them for ongoing processing (as opposed to the previous method of
just rereading the old spec files and using them to recreate the
"STodoTarget"s - i.e., using the same logic as when initially creating
"STodoTarget"s from a new spec).

Changed the data structure used to hold the "targets" (all of the
STodoTarget descendants) from Array to Hash, with the "handle" as the key.
(The "targets" hash table is what is saved via serialization.) Also, the
file of "stored handles", used to ensure that no new spec with a duplicate
handle is accepted, is no longer used, since "targets", with the handle
hash key, can now be used for this purpose.

Added two new classes: AssertionError and PreconditionError (child of
AssertionError), which are "RuntimeError"s used to throw code-assertion
errors.  Used PreconditionError to code a precondition for
perform_initial_processing.