ongoing vs initial emails, bug fix
Changed the scheme for specifying whether an email address for an s-todo
target is to be used for ongoing vs initial email notifications - i.e.,
initial-only is specified with <name>@<host>.com[initial] and ongoing-only
with <name>@<host>.com[ongoing]; to specify that an address is to be used
for both don't include a tag: <name>@<host>.com.

Bug fix: although duplicate handles (which need to be unique) are prevented
between any new specs and existing (old) specs/s-todo-targets, duplicate
handles among new specs encountered in the same run were not being
detected.  Since the resulting "STodoTarget"s (newly created from the new
specs) were being stored in a hash table with the handle as the key,
all new "STodoTarget"s with the same handle were being "pruned" simply
according to the order in which the spec files were being read.  The fix
was to look for duplicate handles in new specs, report them, and prevent
creation (and storing) of any "STodoTarget"s using the conflicting handle
(leaving the spec files using that handle in the "specpath" directory)
so that the user can fix the problem and rerun the application.

As part of making the change re. ongoing vs initial emails, changed
STodoTarget: Made the 'initial_email_addrs' and 'ongoing_email_addrs'
fields public (via attr_reader) and changed its 'initialize' method to set
both of these attributes, from the 'email_spec' (renamed from 'email')
obtained from the spec file, upon object creation.