Began implementation of reminder processing.
Created a Reminder class.  "STodoTarget"s now have a set of (0 or more)
Reminder instances based on the associated specification.  These reminders
will be used (via their date_time attribute) to determine if it is time to
trigger a reminder (send email or whatever) when a Reminder's datetime is
the same (or very close to) the current time.

Added some validation logic based on an STodoTarget's reminders.  At
construction time for an STodoTarget, if any of the reminders has a bad date
(i.e., specified in the spec file for that target; e.g., '2018/02/30'), no
corresponding Reminder object is created, processing of reminders for that
target is abandoned, and the STodoTarget is created with 'valid?' being
false.  Any invalid "STodoTarget"s are excluded from the list of targets
(logic residing in TargetBuilder) and not processed; the corresponding spec
file is left as is.  This is to allow the user to fix the problem in the spec
file and allow it to be resubmitted.

Cleanup: Removed a bit of old, obsolete code.