CompositeTask.descendants, etc.
Added a 'descendants' query to CompositeTask, with related changes in that
class and STodoTarget.  Used the new query in
STodoManager.report_descendants.  In STodoManager, don't allow a child to
be added to a non-CompositeTask (checkable via the new can_have_children?
query).  Moved the 'parent_handle' attribute from CompositeTask up to
STodoTarget.