bug fix re. DateTime -> Time
Use of DateTime (with 'parse' and 'strftime') for datetimes was resulting in
UTC datetimes instead of local time. (E.g., DateTime.parse('2016/10/22 18:00')
created a datetime with local time as 23:00.) This was showing up in
the Reminder times by triggering reminders a few hours too early. It was
probably also causing other problems I'd not noticed. Fixed by switching to
the Time class.

CalendarEntry: Removed an old, renamed version of entry_creation_command
that I forgot to delete earlier.