spec templates, bug fix, etc.
Added a 'final_reminder' query to the STodoTarget descendants and used it
in the code to send notifications (emails) to the user, so that, e.g., a
reminder will be sent to the user when a task's due-date or an
appointment's date-time passes.

Some improvements to the ErrorTools facility to make it easier to code
assertions and provide more info when one is triggered.

Bug fix: When a spec file is archived, if the name conflicts with a file
already in the archive directory, rename it to prevent overwriting.

Added functionality, and a driver script, to output a template spec file
based on a user-provided type argument (task, memo, ...).

Created 'stodo', a high-level driver script to provide the commands:
new|init, notify, report, ...