Switched to notifier model for email.
Changed the design for sending email notifications:  Changed to a model in
which STodoTarget uses 'notifiers' to send a notification instead of
directly knowing about and using Email objects.  The goal of this change is
to allow for easy addition of other notification methods (the most obvious
probably being text messages) with little or no change to the core
STodoTarget class.  Instead of setting the subject, body, etc. in an Email
object, STodoTarget now sets its own fields (notification_subject,
full_notification_message, ...) to be used by the notifier as needed (E.g.,
Email will grab 'notification_email_addrs' while a TextMessage would grab
[to-be-added-if/when-needed] 'notification_phone_numbers'.) and then passes
it"self" to the notifier's 'send' method.  Of course, since STodoTarget has a
list of notifiers, it allows for more than one notification method to be
used for a notification.

Added 'assert' method to ErrorTools.