CalendarEntry improvements;configurability changes
Some changes/improvements to the logic in CalendarEntry for creating a
calendar entry (currently using gcalcli); the main goal is better error
reporting when something goes wrong, such as the gcalcli command failing.

Made the calendar command (calendarcmd) configurable via the config file.

Changed the Logger initialization to use a file instead of stderr or
stdout.  The path/name of this file is overridable with an environment
variable.