Organized into subdirectories.
Added a bit more organization to the system by moving ruby source files
into appropriate subdirectories.  Changed the driver 'stodo' to load the
paths of these subdirectories into the RUBYLIB environment variable so that
ruby will be able to find them.  Corresponding with that change, changed
all occurrences of require_relative to require.