Improved subject line for ongoing messages
In 'send_ongoing_notifications', changed the constructed subject line
such that the date/time is at the end and (by replacing "Reminder:" with
"todo") made sure the 'current_message_subject' is about as far to the
left on the subject line as it can be.  The goal, of course, is to allow
the user to see the most pertinent subject information, since many email
interfaces don't show the entire subject line in their summary view.