Implemented deletion of targets.
Added del[ete] option to stodo to delete STodoTarget objects and added the
necessary code to implement it, including new driver script,
stodo-mgmt.rb, and new class STodoTargetEditor, as well as needed additions
to STodoManager and CompositeTask.  The target to be deleted is removed
from the Hash of "STodoTarget"s and, if it is a child of a CompositeTask,
removed from the parent's list of children.