Implemented editing of "target"s.
"STodoTarget"s are edited by running initial-processing.rb (stodo init)
with a spec file in the spec directory with a type of (instead of task,
appointment, note, etc.) "correction", which instructs the system to,
instead of creating a new STodoTarget of the specified type, find the
STodoTarget (in the 'existing_targets') with matching handle and, for any
fields specified in the spec file, set each field to the specified value.
(Fields not to be changed, of course, should be excluded from the file.)

For this new functionality, a design change was made to TargetBuilder's
interface: the new public method build_targets now needs to be called in
order for new targets to be built and/or, with the new editing feature,
existing targets edited.  'build_targets' is called by STodoManager in its
constructor.

In TargetBuilder, changed @@target_factory_for to @target_factory_for
(i.e., class variable to instance variable), which is now initialized at
creation time.

Added 'modify_fields' method to STodoTarget and its descendants as part of
target editing/correction implementation.