Improved reminder reporting.
When only the first upcoming reminder is being reported (i.e., 'stodo report
remind', not 'stodo report remindall'), display the reminder-date/time
first - at the beginning of each line so that report is lined up in a table-
like format; and report the title instead of the handle (since it usually
gives more info about the task/appointment/...).  Also, for "STodoTarget"s
for which the latest reminder is in the past, display that latest reminder
(instead of the [nonexistent] first upcoming reminder), but put it in
parentheses to make it easy to spot that the date has already passed by.
Note that "STodoTarget"s with no reminders are not reported.