Report due dates.
New reporting feature: List items (currently with title, type, and handle)
with their due date (i.e., date_time for ScheduledEvent, expiration_date
for Memorandum, and due_date for CompositeTask), sorted by due date.  If
the due date is in the past, surround it with parentheses.  Note: The
'<=>' operator was added to STodoTarget so that an array of
"STodoTarget"s (tasks, memos, etc.) can be sorted easily by 'time'.