summary report improvements, spec files
Enhanced the summary report (stodo report summary [handle ...]) to, if one
or more handles are specified, include information about task descendants
if there are any.  To implement this, ReportManager.target_info (utility
routine) was changed to use the TreeNode class and its descendants_report
method.  To accommodate the behavior needed by 'target_info',
'descendants_report' was changed to take a "cutoff" argument.
'target_info' calls it with a cutoff of 1 (in the case where no handles
were specified - i.e., a report on all targets) to indicate to only report
on the root (top ancestor), and with a cutoff of -1 (when 1 or more handles
were specified) to indicate to report on all descendants with no cutoff.
(For mathematical completeness, a cutoff of 0 indicates an empty report.)
Also, fixed a bug in TreeNode.children that showed up in testing these
changes.

Also, added some fake/example spec files to give users some sense of how to
create these files; and added an example 'config' file.