Marshal -> YAML::Store, etc.
For persistence of STodoTarget (and descendants) objects, changed from the
Marshal library to YAML::Store.  Summary of changes needed for this switch:

  - A new class, YamlStoreBasedDataManager, was created.  It is
    created/initialized in Configuration (instead of FileBasedDataManager)
    as @data_manager.
  - A 'user' name/id was added (configurable or else taken from $USER).
    This is used as the main key for storage - e.g., store[user] = targets.

Note that the 'marshal_dump' and 'marshal_load' implementations (used by
the Marshal library) were left in.  If the switch to YAML::Store becomes
permanent, they should be removed.

Also, a couple small reporting improvements were made, such as use of the
new time_24hour method for sane date/time display.  And added the concept
of 'dirty' (boolean) in STodoManager, which is used in
'perform_ongoing_processing' via this call: t.perform_ongoing_actions(self)
t is a STodoTarget - it sets the 'dirty' attribute of 'self'
(status_client) to true if it makes any changes to its persistent
attributes (namely, its Reminders) in 'perform_ongoing_actions'.  After
t.perform_ongoing_actions is called, of course, the STodoManager saves the
data (all 'existing_targets', actually) if 'dirty' is now true (i.e., one
or more of the existing_targets has been changed).