persistence: needed data only, children:Set, etc.
STodoTarget: Implemented new method, prepare_for_db_write, which sets
attributes that don't need to be stored to "empty" values (such as:
'@notifiers = []' and '@email_spec = ""').  Added a call to
prepare_for_db_write in YamlStoreBasedDataManager.store_targets to make use
of this method to avoid saving unneeded data to the persistent-store file.
(Also in YamlStoreBasedDataManager, added "@store.ultra_safe = true".)

Changed name of marshal_dump and marshal_load methods in STodoTarget and
descendants to remind myself that these methods are obsolete and should
soon be removed.

ReportManager.report_complete: Added list (handles) of tasks/children.

Bug fix: Memorandum and ScheduledEvent objects should be allowed to be
children to a specified parent.  A 'correction' to add one of these objects
as the child of a parent object was failing.  Fixed by moving these lines:
    if spec.parent then
      @parent_handle = spec.parent
    end
from CompositeTask to STodoTarget, so that all STodoTarget can be children.

CompositeTask: Changed the '@tasks' (child targets) attribute type from
an Array to a Set, to enforce that a child should only occur in its parents
list of children/tasks once.