Preparation for redesign: STodoTarget as composite
In preparation for moving the composite quality of CompositeTask to its
parent, STodoTarget (at which point CompositeTask will, of course, have its
name changed to Task), changed the name of the 'tasks' attribute in
CompositeTask to 'children'.  The main point of this change is that when
the complete change is implemented, STodoTarget objects will have
sub-objects that could be tasks or could be memos (AKA notes) or
appointments (AKA ScheduledEvents), so in preparation for this change, the
'tasks' attribute (originally representing a set of child tasks) has been
renamed 'children' to reflect that the objects in this set can be of a type
other than Task (or CompositeTask).  As well as the 'tasks' -> 'children'
change, it was also appropriate to change 'task' to 'child' - e.g., add_task
was changed to add_child.