Added states: in-progress, completed, etc.
Added a 'state' attribute, of type TargetState (new class) to STodoTarget.
Possible states are in-progress, suspended, canceled, and completed.  Made
use of the state in several places, such as in STodoTarget.to_s and
ReportManager.report_due (stodo report due), in which targets with a state
of canceled or completed are now excluded.

Implemented changing the state of existing targets via
'stodo stat <state-change> <handle> ...' (example: stodo stat suspend foo1).
Implemented STodoTargetEditor.modify_state to help implement this
state-changing functionality and added it to the @method_for hash (i.e.,
with 'state' as key and modify_state as value).  Invalid state transitions
(such as completed -> in-progress) are guarded against and a warning is
output if such a state change is requested.