Implemented backup functionality.
Implementation providing backup of persistent data: Created new main
script, stodo-admin.rb, which oversees administrative functionality
(currently only backup) and a new class: STodoAdministrator, which, of
course, carries out administrative functionality delegated to it by
stodo-admin.rb.  It essentially mimics the design of the STodoTargetEditor
class in that it invokes the requested command via a table
of methods, so that the 'backup' request is mapped to the
'perform_data_backup' method (with the code: 'method = @method_for[command]'
and, a bit later, 'self.send(method, command_with_args[1..-1])').

Since the backup implementation, rather than simply copying the persistent
data file, uses the YAML::Store facility to read the application's
persistent (YAML::Store) file and (if the new 'last_update' [Time]
value in the main file/database is later than that of the backup database)
write the resulting contents to the backup file/database, the
YamlStoreBasedDataManager class was extended to provide this
functionality.  Thus the main backup logic is carried out by
YamlStoreBasedDataManager.backup_database and (in the same class)
helper methods.

Some backup-related changes were also needed to Configuration (i.e., read
the new backuppath setting and make it available to Configuration clients)
and the stodo script - to provide the user-facing interface for backups -
i.e., 'stodo backup'.

Unrelated to backups, also changed the name of the 'send' method to
'send_message' (in Email and Mailer) to prevent any possible future
confusion with the well-known Object.send method.