TargetStateSet, template improvements
New TargetStateSet abstraction to provide a set of states to use to specify
in reports which STodoTarget states (e.g., "only in-progress" or "canceled
and completed") to include in the report.  (Example command line:
stodo report due:inprog,susp)  Modified ReportManager and report.rb to use
this new class.

For templates (stodo temp <type>), added a 'template = false' argument to
to_s in STodoTarget.  When true, the output is in a format expected for
spec files.  (The default [template = false] uses slightly different
format.)  Also added (optional) options-with-accompanying values for the
'stodo temp' command to allow the user to specify the fields to supply for
a new specfile. When the -h option is used (e.g., 'stodo temp action -h foo'),
the result will be (once a bug in this new feature is fixed) a file that
can be used as is as a stodo specfile, although usually more fields will be
filled in via editing the result and/or supplying more options.  ('stodo
help temp' shows all available options.)