2 bug fixes: template & multi-line handle
STodoTarget.to_s, when template == true, was including the status, which
does not occur in a spec file and thus should not be output in 'template'
mode - the fix was simple:
-          result += "status: #{state}\n"
+          if ! template then
+            result += "status: #{state}\n"
+          end

STodoSpec.extract_settings: When a handle is specified that spans more than
one line, the multi-line handle was accepted and stored.  But multi-line
handles should not be allowed, since they cause problems.  (For one, how
does one specify a multi-line handle for a report?  There's no way to do it
from the command line.)  The fix was to discard, in the handle value, the
first newline and all characters that follow that newline.  I.e.:
         key = standardized_key(components[i].sub(/: */, ""))
         value = components[i + 1]
+        if key == HANDLE_KEY then
+          value = value.sub(/(?m:\n.*)/, "")
+        end