improvements to "stodo temp" and reporting
Implemented a method of, for the 'stodo temp' (spec-file templating)
option, installing the new spec file to the 'spec_path' directory without
committing it to be processed by inserting the string 'type: template'
into the spec file.  The user edits this file (when the editor [$EDITOR env.
var] is invoked via 'stodo temp') and, to allow the file to be processed,
removes that line before writing the file and exiting the editor.  If the
'type: template' line remains in the file (below the earlier, valid
'type: ...' line) the 'stodo init' command will ignore the spec file, leaving
it in place.  The user can then later remove the 'type: template' line to
allow the file to be processed - i.e., used to create a new STodoTarget.

Changed the reporting (report.rb) logic such that if no states are specified
by the user, only active "STodoTarget"s are reported.  For example,
'stodo report due' lists only items that are in-progress or suspended, not
canceled or completed - to list all items, one can run: 'stodo report due:e'
(which matches all states, since the letter e occurs in all four state
names).  The assumption here is that the user will not usually need to know
about "STodoTarget"s that are canceled or completed.

.stodo_utilities: Added a couple more convenience aliases.