small improvements: reminders spec, stodo temp
Since reminders can contain more than one date and since a date can have a
comma in it (e.g., 'jan 2, 2017'), changed the delimiter for reminders to
';' (semi-colon) to allow several dates that contain commas.

stodo: make_template: Added error check for invalid type, which occurs when
the type is not specified and the argument in its place is a flag
(e.g., 'stodo temp -t <title> ...').