Improved date parsing
Extended the date parsing facilities to use an external program (a perl
program that uses Date::Manip) to handle date/time specifications that
Time.parse cannot handle, with the help of 2 new classes, DateParser and
ExternalDateParser (which calls dates.pl).  This new mechanism is currently
only employed for "Reminder"s.  It needs to be used also for the date/time
(due_date, date_time, expiration_date) attribute of the STodoTarget
descendants (Task, Memorandum, ...).  It also needs to be made more robust
in the way it is used for "Reminder"s.