Bug fix: Appointment template
Appointment templating was failing - @valid was being set to false because
there was no date, which for a "template" is expected.  Changed the logic
to set @valid to spec.is_template? (new query in STodoSpec and StubbedSpec)
so that @valid will always be true when a "template" spec (i.e.,
StubbedSpec) is being used.