Refactoring of reminders/periodic-reminders
Refactored the reminder logic into a proper design: renamed Reminder to
OneTimeReminder; created a new Reminder class that became parent of
OneTimeReminder and PeriodicReminder, along with the necessary moving of
common features up to the new parent, etc.; changed code that was
referencing Reminder to OneTimeReminder (e.g.:
-        @final_reminder = Reminder.new(expiration_date)
+        @final_reminder = OneTimeReminder.new(expiration_date)
).  Also, removed the concept of 'triggered?' in Reminder (the parent) -
the triggered? query now only occurs in OneTimeReminder.  (It doesn't make
much sense for a PeriodicReminder to be either triggered or not triggered.)
Also removed debugging code ($log.debug..., etc) that was added to
develop/test the PeriodicReminder logic, as well as the PeriodicDateParser
logic.

STodoTarget.set_initial_calendar_entry: Added a guard for the
calentry.submit call to prevent the call if calentry's duration, title, or
time values are nil.

.stodo_utilities: small bug fix re 'stodo temp ...' (in tempm, tempt, ...)
time argument.