ongoing notification - reminders/date fix
STodoTarget.send_ongoing_notifications:  Now that 'upcoming_reminders' are
part of the report/message, r.trigger (Reminder.trigger) needs to be
called before constructing the message.  Otherwise, there won't be any
upcoming reminders because the STodoTarget's reminders will still have a
date-time in the past (since 'trigger' was not called yet), meaning that
they are due (is_due?) and thus are not upcoming.  But the
reminder-date-time for the report needs to be grabbed from the reminders
before 'trigger' is called, since the current datetime is needed for the
report, not the upcoming datetime.