System/log file handling improvements
STodoTarget.set_initial_calendar_entry: Removed the check for
"calentry.duration != nil" as a condition for calling "calentry.submit".
The CalendarEntry will, in its 'entry_creation_command' method (called by
"submit") set the duration to 1 if it is currently nil - that is, it will
cause the creation of a calendar entry with a very small duration (1
minute).  Such a small duration can make sense for tasks and memos - e.g.,
an entry can be created for the due date of a task, for which no duration
is needed.  Since a duration is required, 1-minute is used to mean no
duration.