category prefix for reports/emails
Implemented the concept of "category prefix", a prefix to be used for
categories in reports.  The prefix is configured in the configuration file.
For example, a configuration of "categoryprefix = cat:", for a task with
categories "music" and "art", will result in emailed reminders for the
task with this string as part of the subject: "cat:music, cat:art".  The
purpose of this facility is to provide more power for use of the filtering
system provided by your email client.

Added an example entry for categoryprefix in the sample configuration
file (doc/config), as well as a couple other entries/configs that were
missing.