Fix: change-of-parent: failed to update old parent
Added a new 'orig_parent' argument to modify_fields and
main_modify_fields in STodoTarget and its descendants to allow
the STodoTarget object to remove itself as one of the old parent's
children. Modified 'edit_target' in TargetBuilder to obtain the
old/original parent from @existing_targets and pass it, as the
new argument, to "t.modify_fields".