Error-handling logic for invalid parent spec
Changes to deal with an invalid parent-handle being specified for a
parent-change operation for an item (STodoTarget descendant). Involved
changes to the following files:

src/core/stodotarget.rb: In 'main_modify_fields', where the main
parent-change logic is implemented, the validity of 'spec.parent' is
checked, and if it is non-empty and is not a key for a member of
'target_list' (note main_modify_fields 2nd argument name and type change),
which in the caller is called @existing_targets, an exception is raised
with an error message that the new parent-handle is invalid.
'main_modify_fields' also implements the new specs that if spec.parent is
is an empty string, self's parent is changed to nil, indicating that self
(a STodoTarget descendant) is a top-level item - i.e., has no parent. (As
with the other fields in spec, if spec.parent is nil, no change is made.)

src/coordination/targetbuilder.rb: 'build_targets' in the TargetBuilder
object was changed to "rescue" the raised exception in the case where
the current spec ('s', within a for loop) has an invalid parent handle
(the raise occurring down the call stack, beginning at 'target_for').
In this case, the raised error-message/exception is logged and the loop
is allowed to continue - i.e., the current invalid spec is discarded
(not added to @targets), and the next spec, s, in the loop is processed.

src/core/memorandum.rb, src/core/project.rb,
src/core/scheduledevent.rb, src/core/task.rb:
Updated to match the 2nd argument name (and type)
change (orig_parent -> target_list) to 'main_modify_fields' in the
ancestor class, STodoTarget.

Other changes:

src/error/errortools.rb b/src/error/errortools.rb: 'implies_eval2' was
added to provide an alternative 'implies' designed to handle some cases
where the 2nd argument, expr2, could trigger a runtime error due to the
expression needing to be evaluated before making the method call. Some,
but not all, of these cases can now be avoided by using this function,
which expects a String as the 2nd argument, which is only "eval"d and
applied if the first argument is true (i.e., '! expr1' is false).

src/stodo b/src/stodo: beginnings of changes for the planned
"change-parent" command/operation.

.env: Put these "Tools/utilities for development and admin of this project"
under version control.