item-editing/building redesign - interim commit
This a moderate (as opposed to large/extensive) redesign of
the functionality for creating and editing STodoTarget objects,
with changes to STodoManager, TargetBuilder, etc., with the purpose
of improving the design a little to make it less muddy - easier to
understand and change. "interim" because although most (according
to current knowledge) of the needed changes have been made, it is
not quite complete, and the code needs to be cleaned up. The reason
for the commit is to save these changes and the current state of the
code before finishing/cleaning it up. It has been tested somewhat -
mostly working correctly except for the new 'change' feature (which
incorrectly changes features that have not been targeted for changing,
such as the email address and the date). And a few, relatively minor
(I think) changes still need to be made.

SUMMARY OF MAIN DESIGN CHANGES:

STodoManager: Several changes: Use git diff against older revisions
to review them. (I don't want to take the time to summarize them here
at this point.)

TargetBuilder: The most important change in this class is probably
moving of the logic of the creating and adding to the 'targets'
array/attribute from the constructor (i.e., "initialize") to separate
methods that are invoked after construction - i.e., by a client calling
'build_targets'. 'prepare_targets', called by 'build_targets' simply
initializes self.targets to an empty array. Again, use 'git diff ...'
to review changes in detail.

TemplateTargetBuilder: Like its parent, TargetBuilder, the creation of
the 'targets' array has been moved out of 'initialize'. In this case,
'prepare_targets' creates 'targets' as an empty array and then
populates it - with one STodoTarget object (since this class deals
with only one STodoTarget, in order to fulfill the originally intended
behavior, which is to create a "template" of an item [a string with the
STodoTarget's attribute values] to be viewed and/or edited, by the
user). This "target" is obtained via a call to 'target_for(spec)', where
spec is an attribute whose value has been set to a new StubbedSpec (which
will often/usually obtain its "specs" from the command-line - ARGV)
object in the constructor. 'prepare_targets' needs to be called from a
client object after the TemplateTargetBuilder has been
created/constructed.

TargetEditor: A new class, which inherits from TemplateTargetBuilder
in order to - instead of generating a template - provide "editing"
functionality from within the program (e.g., by carrying out a command
implied by command-line options provided by the user). Since, from
the class's "point of view", the functionality is so similar to that
of TemplateTargetBuilder (TargetEditor itself contains very little
code), it might be feasible to adapt TemplateTargetBuilder to provide
the "editing" functionality and get rid of this class, in which case
TemplateTargetBuilder should, perhaps, be renamed.

src/main/stodo-mgmt.rb, src/main/stodo-template.rb: Changes needed to
adapt to the new scheme with respect to TargetBuilder (and its
descendants), STodoManager, etc.

MISCELLANEOUS SMALL CHANGES WORTH MENTIONING:

Changed all occurrences of the File.exists? query to the "new"
convention: File.exist?. Due to switching to ruby 3 for most testing,
the "exists?" version is obsolete and has been removed.

FileBasedSpecGatherer: Added 'public', above attribute :specs, to
make the "publicness" explicit. Added inclusion of Contracts::DSL.
(Added method 'spec_for', which is not used and will probably be
removed.)

src/stodo: Small changes to "help" messages.