stodo change ... bug fix; new: re-adopt... feature
A bug caused the "stodo change ..." command to change fields that should
not have been changed; for example, the email field was set to its default
value (which is, at this time, <template>@<template.org>) when no email
value was specified on the command line (i.e., no option of '-e <addr>');
another example: when no -d (description) option was given, the description
value for the specified handle/item was changed to be empty.

TemplateOptions: For the new changes to work correctly, the attributes
of TemplateOptions need to be initialized to nil rather than an empty
String.

StubbedSpec: Added 'use_defaults' argument (which defaults to true) to
the constructor - if it's false, the default values assigned (in
'set_fields') to handle, email, due_time, and expire_date are not used
for initialization.

TemplateTargetBuilder: Changed constructor to take an optional
'the_spec' argument and if it is not nil, use that as the 'spec'
attribute instead of creating a new StubbedSpec object - i.e.:
-  def initialize options, the_existing_targets = nil
+  def initialize(options, the_existing_targets = nil, the_spec = nil)
     self.existing_targets = the_existing_targets
-    self.spec = StubbedSpec.new options
+    if the_spec.nil? then
+    # (Initialize self.spec with defaults [2nd arg]:)
+      self.spec = StubbedSpec.new(options, true)
+    else
+      self.spec = the_spec
+    end
     super nil

The only top-level code that needed to change as a result of the
StubbedSpec and TemplateTargetBuilder changes described above was for
the 'when /change*/' case in stodo-mgmt.rb - i.e.:
     $log.debug "arguments: #{arguments}"
-    target_editor = TemplateTargetBuilder.new(
-      TemplateOptions.new(arguments, true), manager.existing_targets)
+    options = TemplateOptions.new(arguments, true)
+    spec = StubbedSpec.new(options, false)  # false => don't use defaults
+    target_editor = TemplateTargetBuilder.new(options,
+                                              manager.existing_targets, spec)
     manager.target_builder = target_editor
     manager.update_targets

STodoTarget: As a main component of the 'stodo change' bug fix, changed
'main_modify_fields' to use a new method, 'guarded_scalar_assignment', to
assign new values to its attributes from the 'spec' argument, the purpose
being to ignore spec "settings" that are either nil or empty and thus only
perform the assignment to an attribute if the corresponding setting from
'spec' is not empty - i.e., a non-empty String or an Integer.
Added 'adopt_descendants' method - for the re-adopt feature.

Here is the help info for the re-adopt feature (from the new 'stodo h admin'
help option):
  readopt <h>...     re-"adopt" emancipated descendants of the items with
                     handles <h>...

test/add-change-delete-etc-101: bug fix