Completed references; started on attachments
Completed implementation of current plan for 'references' functionality:
  o STodoTarget 'references' attribute
  o adding and changing 'references' list for an item.
  o checking reference handles and, if STODO_REJECT_BADREFS env. var. is
    set and the associated STodoTarget doesn't exist, rejecting them.

Completed implementation of initial plan for 'attachments' functionality.
Most of the 'attachments' implementation mirrors that of the 'references'
feature:
  o STodoTarget 'attachments' attribute
  o adding and changing 'attachments' list for an item.
  o some "infrastructure" for "rejecting" non-existent attachment files,
    via the STODO_REJECT_BADATTCHMTS env. var.

Notes:
In TargetBuilder, in 'process_targets', this line is crucial:
    s.existing_targets = self.existing_targets
since 's' (a STodoSpec resulting from the loop: for s in self.specs do)
now has a 'existing_targets' attribute, which needs to be available for
checking of references.

In src/main/stodo-mgmt.rb, this change was made:
@@ -24,7 +24,7 @@ if ARGV.length > 1 then
     require 'templatetargetbuilder'
     require 'templateoptions'
     target_builder = TemplateTargetBuilder.new(
-      TemplateOptions.new(arguments, true))
+      TemplateOptions.new(arguments, true), manager.existing_targets)
     target_builder.set_processing_mode TemplateTargetBuilder::CREATE_MODE
     manager.target_builder = target_builder
     manager.add_new_targets
since the TemplateTargetBuilder needs the 'existing_targets' to be set
before its 'process_targets' method (called by manager.add_new_targets)
is called.

There are two new tests: test/attachments-101 and test/references-101.