Started: viewing/editing of attachments
Added new 'proca' (process attachments for the specified items/handles)
command to the 'stodo' script. It pre-processes the user-supplied
arguments and feeds the result to report.rb. In report.rb, the
'when /^attach/' case handles this 'process attachments' command,
reporter.report_attachments(criteria). report_attachments (method of
ReportManager) determines whether we are "editing" or not ("viewing")
via a new configuration setting and then calls 'process_attachments'
on each of the targets (selected by user's handle-list/criteria).

'process_attachments' (STodoTarget) simply calls 'process' on each
of its attachments. Attachment.process creates a MediaHandler and
delegates the "processing" (viewing or editing) of the attachment to
the handler. The new MediaHandler class uses the new FileCommand
class to determine the file type (via UNIX 'file' command) and invoke
the appropriate external program (currently stubbed), which is specified
based on the file type according to newly added user-config settings.

Other highlights:

Changed the architecture of the 'config/configuration' object such
that it is globally available via the new Configuration.config class
method (which returns Configuration's new @@config class variable,
which is set to itself [self]), allowing collaborating objects to
easily access this 'config' without having to pass a 'config' around
"everywhere". Made appropriate changes/adjustments to this change -
especially in getting rid of no-longer-needed 'config' arguments to
methods.

To-do: Implementation of invocation of appropriate external command to
view or edit the attachment/file.